% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{clean_tmp}
\alias{clean_tmp}
\title{Clean temporary files}
\usage{
clean_tmp(tmpdir = tempdir())
}
\arguments{
\item{tmpdir}{the path to temporary directory}
}
\value{
cleaned temporary directory
}
\description{
Performs clean operations on temporary directory
}
\details{
\code{clean_tmp()} will delete all files stored in temporary
         directory. All extensioned files are remove, but not directories
         inside \code{tempdir()}.
}
\examples{
clean_tmp()

}
\seealso{
\itemize{
         \item{\code{tempfile()}}
         \item{\code{tempdir()}}
         }
}
\author{
Bruno Pinheiro
}
