% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{dark_map_theme}
\alias{dark_map_theme}
\title{Dark theme for ggplot2}
\usage{
dark_map_theme(base_font = "Fira Sans Condensed",
  title_font = "Fira Sans Condensed Medium")
}
\arguments{
\item{base_font}{a string indicating which font to use (default is Fira
Sans Condensed)}

\item{title_font}{a string indicating which font to use (default is Fira
Sans Condensed Medium)}
}
\value{
a predefined \code{ggplot2} theme based on
\code{dark_theme_minimal()} from \code{ggdark}
}
\description{
Implement a dark ggplot2 theme
}
\examples{
ggplot(sample_data) +
  geom_bar(aes(x = var5)) +
  dark_theme()

}
