% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_ibge()}
\alias{get_ibge()}
\alias{get_ibge}
\title{Get tabular data from IBGE}
\usage{
get_ibge(url)
}
\arguments{
\item{url}{a string defining url of zipped shapfile to get}
}
\value{
extrated files in temporary directory
}
\description{
Download and extract tabular data from ftp://ftp.ibge.gov.br/
}
\details{
\code{get_ibge} will download and extract files in a
         \code{tempdir()}. Files keep stored temporarily. To get other
         \code{.zip} is needed to remove files from \code{tempdir()}: see
         \code{\link{clean_tmp}}.
}
\examples{
u1 <- "ftp://geoftp.ibge.gov.br/organizacao_do_territorio/malhas_territoriais/"
u2 <- "malhas_de_setores_censitarios__divisoes_intramunicipais/censo_2010/"
u3 <- "setores_censitarios_shp/al/al_distritos.zip"
get_ibge(paste0(u1, u2, u3))
clean_tmp()

}
\author{
Bruno Pinheiro
}
