% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{light_map_theme}
\alias{light_map_theme}
\title{Light theme for ggplot2}
\usage{
light_map_theme(base_font = "Fira Sans Condensed",
  title_font = "Fira Sans Condensed Medium", base_size = 12)
}
\arguments{
\item{base_font}{a string indicating which font to use (default is Fira
Sans Condensed)}

\item{title_font}{a string indicating which font to use (default is Fira
Sans Condensed Medium)}

\item{base_size}{a numerica value indicating the base font size of the plot}
}
\value{
a predefined \code{ggplot2} theme based on \code{theme_minimal()}
}
\description{
Implement a light ggplot2 theme
}
\examples{
ggplot(sample_data) +
  geom_bar(aes(x = var5)) +
  light_theme()

}
