% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{light_theme}
\alias{light_theme}
\title{Light theme for ggplot2}
\usage{
light_theme(base_font = "Fira Sans Condensed",
  title_font = "Fira Sans Condensed Medium")
}
\arguments{
\item{base_font}{a string indicating which font to use (default is Fira
Sans Condensed)}

\item{title_font}{a string indicating which font to use (default is Fira
Sans Condensed Medium)}
}
\value{
a predefined \code{ggplot2} theme based on \code{theme_minimal()}
}
\description{
Implement a light ggplot2 theme
}
\examples{
ggplot(sample_data) +
  geom_bar(aes(x = var5)) +
  light_theme()

}
