% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{list_tmp}
\alias{list_tmp}
\title{Clean temporary files}
\usage{
list_tmp(ext = NULL, full = TRUE)
}
\arguments{
\item{ext}{a string defining a file extension (defaul is NULL)}

\item{full}{a TRUE/FALSE value indicating if the function should return the
full path of the files (default is TRUE).}
}
\value{
list files in \code{tempdir()}
}
\description{
List files in \code{tempdir()}
}
\details{
\code{list_tmp()} will check and print files in \code{tempdir()}.
         If any file extension is informed, will print all files in the root
         of temporary directory. If some file extension is informed will
         check and print all files matching that extension in
         \code{tempdir()}, including nested folders.
}
\examples{

# to list files in tempdir() root
list_tmp()

# to list files of an extension
list_tmp(".zip")

}
\seealso{
\itemize{
         \item{\code{tempfile()}}
         \item{\code{tempdir()}}
         }
}
\author{
Bruno Pinheiro
}
