% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_bar}
\alias{plot_bar}
\title{Bar plot for univariate analysis}
\usage{
plot_bar(df, x, title = NULL, subtitle = NULL, theme = "light",
  mode = "count", fill = FALSE, flip = FALSE, fillc = "darkgrey",
  base_font = "Fira Sans Condensed",
  title_font = "Fira Sans Condensed Medium")
}
\arguments{
\item{df}{a data frame}

\item{x}{a string indicating a categorical variable}

\item{title}{a string defining the plot title}

\item{subtitle}{a string defining de subtitle of plot}

\item{theme}{a string defining theme: "light" (default) or "dark"}

\item{mode}{a string defining mode: "count" (default) or "prop"}

\item{fill}{a string indicating a categorical variable}

\item{flip}{a logical TRUE or FALSE}

\item{fillc}{bar colors when fill == FALSE}

\item{base_font}{a string indicating which font to use (default is Fira
Sans Condensed)}

\item{title_font}{a string indicating which font to use (default is Fira
Sans Condensed Medium)}
}
\value{
Um grafico de barras para a variavel \code{x} da base \code{df}.
}
\description{
It counts a categorical variable and plots a publication
             ready ggplot2 barplot
}
\details{
Uses dplyr and ggplot
}
\examples{
# simple barplot
plot_bar(sample_data, "var5")

# plot more than one variable at once
vars <- c("var5", "year")
plots <- lapply(vars, function(i) plot_bar(sample_data, i))

}
\author{
Bruno Pinheiro
}
