% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_bar_id}
\alias{plot_bar_id}
\title{Bar plot for univariate analysis}
\usage{
plot_bar_id(df, x, y, order = FALSE, theme = "light", mode = "count",
  fill = FALSE, flip = FALSE, fillc = "darkgrey",
  base_font = "Fira Sans Condensed",
  title_font = "Fira Sans Condensed Medium", title = NULL,
  subtitle = NULL)
}
\arguments{
\item{df}{a data frame}

\item{x}{a string indicating a categorical variable}

\item{y}{a string indicating a numeric variable}

\item{order}{TRUE/FALSE indicating if bars shoul be ordered (default is TRUE)}

\item{theme}{a string defining theme: "light" (default) or "dark"}

\item{mode}{a string defining mode: "count" (default) or "prop"}

\item{fill}{a string indicating a categorical variable}

\item{flip}{a logical TRUE or FALSE}

\item{fillc}{bar colors when fill == FALSE}

\item{base_font}{a string indicating which font to use (default is Fira
Sans Condensed)}

\item{title_font}{a string indicating which font to use (default is Fira
Sans Condensed Medium)}

\item{title}{a string defining the plot title}

\item{subtitle}{a string defining de subtitle of plot}
}
\value{
Um grafico de barras para a variavel \code{x} da base \code{df}.
}
\description{
It counts a categorical variable and plots a publication
             ready ggplot2 barplot
}
\details{
Uses dplyr and ggplot
}
\examples{
library(dplyr)
sample_data \%>\%
  filter(!is.na(var5)) \%>\%
  count(var5) \%>\%
  plot_bar_id(x = "var5",
              y = "n",
              fill = TRUE)

}
\author{
Bruno Pinheiro
}
