% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_na}
\alias{plot_na}
\title{data visualization}
\usage{
plot_na(df, varcut = NULL, title = NULL, theme = "light",
  colors = c("darkgrey", "red"), base_font = "Fira Sans Condensed",
  title_font = "Fira Sans Condensed Medium")
}
\arguments{
\item{df}{a data.frame object}

\item{varcut}{a categorical variable to facet the plot.}

\item{title}{a string defining the plot title}

\item{theme}{a string defining theme: "light" (default) or "dark"}

\item{colors}{a character vector of lenght 2, indicating NA and non-NA
colors in the plot}

\item{base_font}{a string indicating which font to use (default is Fira
Sans Condensed)}

\item{title_font}{a string indicating which font to use (default is Fira
Sans Condensed Medium)}
}
\value{
If \code{varcut} is not passed, then the function plot a simple
        stacked barplot showing the counts of NAs for each variable in
        the dataset. If is, then the counts are faceted.
}
\description{
\code{plot_na} uses dplyr and ggplot2 to return a stacked
             barplot showing the counts of NAs for each variable in the
             dataset.
}
\details{
This function is tidyverse based. It transforms the data in a
         \code{\link{tibble}}, \code{\link{gather}} and \code{\link{count}}
         NAs for each variable. Results are ploted in a stacked bar plot
         with \code{\link{geom_bar}}. Users can define title, colors, and
         can plot results grouped by one variable. In this case the plot
         will be faceted with \code{\link{facet_wrap}} horizontally
         displayed.

         * \code{mode} options: "light" (default), or "dark"
}
\examples{
# plot_na without varcut
plot_na(sample_data)

# plot_na with varcut (useful for time series data)
plot_na(sample_data, varcut = "year")

# plot_na manipulating other parameters
plot_na(
  sample_data,
  varcut = "year",
  title = "Counts of NAs",
  theme = "dark",
  colors = c("red", "darkgreen")
  )

}
