% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{read_geoibge()}
\alias{read_geoibge()}
\alias{read_geoibge}
\title{Get spatial data from IBGE}
\usage{
read_geoibge(layer, epsg = 31983)
}
\arguments{
\item{layer}{a string defining url of zipped shapfile to get}

\item{epsg}{a numeric value defining projection epsg code (default is 31983)}
}
\value{
extrated files in temporary directory
}
\description{
Download, extract and import zipped shapefiles from
             ftp://geoftp.ibge.gov.br/
}
\details{
\code{read_geoibge} will download and extract files in a temp
         directory. Files keep stored temporarily. To remove files from
         \code{tmpdir}, see \code{\link{clean_tmp}}.
}
\examples{
u1 <- "ftp://geoftp.ibge.gov.br/organizacao_do_territorio/malhas_territoriais/"
u2 <- "malhas_de_setores_censitarios__divisoes_intramunicipais/censo_2010/"
u3 <- "setores_censitarios_shp/al/al_municipios.zip"
get_ibge(paste0(u1, u2, u3))
layer <- list_tmp(".shp", full = FALSE)
layer <- gsub("\\\\.shp", "", layer)
st <- read_geoibge(layer)
clean_tmp()

}
\author{
Bruno Pinheiro
}
