% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{read_ibge()}
\alias{read_ibge()}
\alias{read_ibge}
\title{Import tabular data from IBGE}
\usage{
read_ibge(file)
}
\arguments{
\item{file}{a string defining the file name to import}
}
\value{
extrated files in temporary directory
}
\description{
Import tabular data (for now just \code{.csv} format) from
             \code{tempdir()}.
}
\details{
\code{read_ibge} will import informed file from \code{tempdir()}
         directory. Files keep stored temporarily. To remove files from
         \code{tmpdir}, see \code{\link{clean_tmp}}.
}
\examples{
u1 <- "ftp://ftp.ibge.gov.br/Censos/Censo_Demografico_2010/"
u2 <- "Resultados_do_Universo/Agregados_por_Setores_Censitarios/RR_20171016.zip"
get_ibge(paste0(u1, u2))
st <-read_ibge("Pessoa07_RR.csv")
clean_tmp()

}
\author{
Bruno Pinheiro
}
