% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample data used in examples}
\format{A data frame with 500 rows and 6 variables:
\describe{
  \item{var1}{numeric variable}
  \item{var2}{numeric variable}
  \item{var3}{numeric variable}
  \item{var4}{numeric variable}
  \item{var5}{categorical variable}
  \item{year}{categorical variable}
}}
\usage{
sample_data
}
\description{
A dataset with 500 observations and 6 variables: 4 numerics and 2
categoricals, to test functions and produce examples.
}
\keyword{datasets}
